requirements_nonconda.txt
google-api-python-client==1.5.5
httplib2==0.9.2
numpy==1.11.2
oauth2client==4.0.0
pandas==0.19.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.6.0
pytz==2016.10
rsa==3.4.2
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
update==0.4.4
uritemplate==3.0.0
